+++ test_sort.py
@@ -2,7 +2,7 @@
 import shutil
 from argparse import Namespace
 from datetime import datetime
-from pathlib import PosixPath
+from pathlib import PosixPath, Path
 from unittest.mock import patch, call, MagicMock
 
 import pytest
@@ -41,37 +41,44 @@ class TestParseArguments:
 
 class TestGetDateFromFile:
     def test_image_basic(self, sorting_pictures):
-        actual = sorting_pictures.get_date_from_filename('IMG_20171022_124203.jpg')
+        actual = sorting_pictures.get_date_from_filename(Path('IMG_20171022_124203.jpg'))
 
         expected = datetime(year=2017, month=10, day=22, hour=12, minute=42, second=3)
 
         assert actual == expected
 
     def test_image_duplicate(self, sorting_pictures):
-        actual = sorting_pictures.get_date_from_filename('IMG_20171104_104157_01.jpg')
+        actual = sorting_pictures.get_date_from_filename(Path('IMG_20171104_104157_01.jpg'))
 
         expected = datetime(year=2017, month=11, day=4, hour=10, minute=41, second=57)
 
         assert actual == expected
 
     def test_screenshot(self, sorting_pictures):
-        actual = sorting_pictures.get_date_from_filename('Screenshot_20171007-143321.png')
+        actual = sorting_pictures.get_date_from_filename(Path('Screenshot_20171007-143321.png'))
 
         expected = datetime(year=2017, month=10, day=7, hour=14, minute=33, second=21)
 
         assert actual == expected
 
     def test_video(self, sorting_pictures):
-        actual = sorting_pictures.get_date_from_filename('VID_20180724_173611.mp4')
+        actual = sorting_pictures.get_date_from_filename(Path('VID_20180724_173611.mp4'))
 
         expected = datetime(year=2018, month=7, day=24, hour=17, minute=36, second=11)
 
         assert actual == expected
 
     def test_failure(self, sorting_pictures):
-        actual = sorting_pictures.get_date_from_filename('metadata.jpg')
+        actual = sorting_pictures.get_date_from_filename(Path('metadata.jpg'))
 
         assert actual is None
+
+    def test_image_tilde(self, sorting_pictures):
+        actual = sorting_pictures.get_date_from_filename(Path('IMG_20171022_124203~2.jpg'))
+
+        expected = datetime(year=2017, month=10, day=22, hour=12, minute=42, second=3)
+
+        assert actual == expected
 
 
 class TestIsFile:
@@ -155,20 +162,32 @@ class TestMoveFile:
 
     def test_file_exists_different_hash(self, sorting_pictures, tmp_path):
         src = tmp_path / 'src'
+        dest = tmp_path / 'dest'
+
+        if src.exists():
+            shutil.rmtree(src)
+        if dest.exists():
+            shutil.rmtree(dest)
+
         src.mkdir(parents=True, exist_ok=True)
         src_file = src / 'metadata.jpg'
-        shutil.copy2('sample-images/metadata.jpg', src_file)
+        dest_file = dest / 'metadata-dest.jpg'
 
-        dest_file = tmp_path / 'dest' / 'metadata-dest.jpg'
+        shutil.copy2('sample-images/metadata.jpg', src_file)
 
         assert not dest_file.exists()
         sorting_pictures.move_file(src_file, dest_file)
         assert dest_file.exists()
 
-        src_file = src / 'no-metadata.jpg'
         shutil.copy2('sample-images/no-metadata.jpg', src_file)
+        assert sorting_pictures.move_file(src_file, dest_file) is True
+        assert (dest_file.parent / ('%s-%d%s' % (dest_file.stem, 1, dest_file.suffix))).exists()
+        assert sorting_pictures.move_file(src_file, dest_file) is True
 
-        assert sorting_pictures.move_file(src_file, dest_file) is False
+        shutil.copy2('sample-images/no-metadata/IMG_20171022_124203_01.jpg', src_file)
+        assert sorting_pictures.move_file(src_file, dest_file) is True
+        assert (dest_file.parent / ('%s-%d%s' % (dest_file.stem, 2, dest_file.suffix))).exists()
+        assert sorting_pictures.move_file(src_file, dest_file) is True
 
     def test_src_file_is_dir(self, sorting_pictures, tmp_path):
         src = tmp_path / 'src'
@@ -233,7 +252,8 @@ class TestSortImages:
                           PosixPath('dest/2018-07/VID_20180724_173611.mp4'),
                           PosixPath('dest/2017-11/IMG_20171104_104157.jpg'),
                           PosixPath('dest/2017-10/IMG_20171007_143321.png'),
-                          PosixPath('dest/2017-10/IMG_20171022_124203.jpg')]
+                          PosixPath('dest/2017-10/IMG_20171022_124203.jpg'),
+                          PosixPath('dest/2017-10/IMG_20171022_124203-1.jpg')]
 
         log = sorting_pictures.log
         log['parse_date'] = [p.relative_to(tmp_path) for p in log['parse_date']]
@@ -242,10 +262,7 @@ class TestSortImages:
             'parse_date': [PosixPath('src/metadata-copy.jpg'), PosixPath('src/no-metadata.jpg'),
                            PosixPath('src/metadata.jpg')],
             'suffix': [],
-            'collisions': [
-                (
-                    PosixPath('src/no-metadata/IMG_20171022_124203_01.jpg'),
-                    PosixPath('dest/2017-10/IMG_20171022_124203.jpg'))]
+            'collisions': []
         }
 
     def test_successful_run_move(self, sorting_pictures, tmp_path):
@@ -268,13 +285,13 @@ class TestSortImages:
                           PosixPath('dest/2018-07/VID_20180724_173611.mp4'),
                           PosixPath('dest/2017-11/IMG_20171104_104157.jpg'),
                           PosixPath('dest/2017-10/IMG_20171007_143321.png'),
-                          PosixPath('dest/2017-10/IMG_20171022_124203.jpg')]
+                          PosixPath('dest/2017-10/IMG_20171022_124203.jpg'),
+                          PosixPath('dest/2017-10/IMG_20171022_124203-1.jpg')]
 
         result = sorting_pictures.search_directory(src)
         result = [p.relative_to(tmp_path) for p in result]
         assert result == [PosixPath('src/metadata-copy.jpg'), PosixPath('src/VID'), PosixPath('src/no-m'),
-                          PosixPath('src/no-metadata.jpg'), PosixPath('src/metadata.jpg'), PosixPath('src/no-metadata'),
-                          PosixPath('src/no-metadata/IMG_20171022_124203_01.jpg')]
+                          PosixPath('src/no-metadata.jpg'), PosixPath('src/metadata.jpg'), PosixPath('src/no-metadata')]
 
         log = sorting_pictures.log
         log['parse_date'] = [p.relative_to(tmp_path) for p in log['parse_date']]
@@ -283,8 +300,7 @@ class TestSortImages:
             'parse_date': [PosixPath('src/metadata-copy.jpg'), PosixPath('src/no-metadata.jpg'),
                            PosixPath('src/metadata.jpg')],
             'suffix': [],
-            'collisions': [(PosixPath(
-                'src/no-metadata/IMG_20171022_124203_01.jpg'), PosixPath('dest/2017-10/IMG_20171022_124203.jpg'))]
+            'collisions': []
         }
 
     def test_unknown_suffix(self, sorting_pictures, tmp_path):