+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2020-02-12
+### Added
+- Added progress bar.
+
 ## [0.7.0] - 2020-02-12
 ### Added
 - Log of files not copied or moved.
@@ -13,7 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Collisions that have a different hash will have an index number appended to
   the name before the suffix in the format `-%d`. For example: `IMG_20200212_090807-1.jpg`.
 - Added tilde (`~`) to replacement list in image filenames.
-- Added progress bar.
 
 ### Changed
 - Refactored into a class.